Shifted InChI functions to separate files
