+++ src/bridge/mod.rs
@@ -4,6 +4,9 @@ pub use descriptors::ffi as descriptors_ffi;
 mod fingerprint;
 pub use fingerprint::ffi as fingerprint_ffi;
 
+mod inchi;
+pub use inchi::ffi as inchi_ffi;
+
 mod mol_standardize;
 pub use mol_standardize::ffi as mol_standardize_ffi;
