+++ wrapper/src/ro_mol.cc
@@ -6,7 +6,6 @@
 #include <GraphMol/Fingerprints/Fingerprints.h>
 #include <GraphMol/MolStandardize/Tautomer.h>
 #include <GraphMol/MolOps.h>
-#include <GraphMol/inchi.h>
 
 namespace RDKit {
     using ExplicitBitVect = ::ExplicitBitVect;
@@ -47,16 +46,5 @@ namespace RDKit {
         }
 
         return std::unique_ptr<std::vector<std::string>>(get_types);
-    }
-
-    rust::String mol_to_inchi(std::shared_ptr<ROMol> mol) {
-        ExtraInchiReturnValues rv;
-        return MolToInchi(*mol, rv, "");
-    }
-
-    std::shared_ptr<ROMol> inchi_to_mol(const std::string &inchi) {
-        ExtraInchiReturnValues rv;
-        ROMol *mol = InchiToMol(inchi, rv);
-        return std::shared_ptr<ROMol>(mol);
     }
 }
